move vmrx_to_concx test
