+++ tests/test_mathutils.py
@@ -2,7 +2,6 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
-from pyaerocom.aux_var_helpers import vmrx_to_concx
 from pyaerocom.mathutils import (
     _nanmean_and_std,
     calc_statistics,
@@ -11,21 +10,6 @@ from pyaerocom.mathutils import (
     is_strictly_monotonic,
     make_binlist,
 )
-
-
-@pytest.mark.parametrize(
-    "inputval,p,T,vmr_unit,mmol_var,mmol_air,to_unit,desired",
-    [
-        (1, 101300, 293, "nmol mol-1", 48, None, "ug m-3", 1.9959),
-        (1, 101300, 273, "nmol mol-1", 48, None, "ug m-3", 2.1421),
-        (1, 101300, 273, "nmol mol-1", 48, None, "kg m-3", 2.1421e-9),
-        (1, 101300, 273, "mol mol-1", 48, None, "kg m-3", 2.1421),
-        (1, 98000, 273, "mol mol-1", 48, None, "kg m-3", 2.0724),
-    ],
-)
-def test_vmrx_to_concx(inputval, p, T, vmr_unit, mmol_var, mmol_air, to_unit, desired):
-    val = vmrx_to_concx(inputval, p, T, vmr_unit, mmol_var, mmol_air, to_unit)
-    assert_allclose(val, desired, rtol=1e-4)
 
 
 @pytest.mark.parametrize("vmin, vmax, num", [(0, 1, 10), (0.345, 0.346, 100), (-2, -10, 5)])
